rename subnet dir
