+++ subnetactor/types.go
